(REF) Extract CacheTrait
